Struct csscolorparser::Color

pub struct Color {
    pub r: f64,
    pub g: f64,
    pub b: f64,
    pub a: f64,
}

The color

Fields
r: f64

Red

g: f64

Green

b: f64

Blue

a: f64

Alpha

Implementations

impl Color

pub fn from_rgb(r: f64, g: f64, b: f64) -> Color

Arguments:

r: Red value [0..1]

g: Green value [0..1]

b: Blue value [0..1]

pub fn from_rgba(r: f64, g: f64, b: f64, a: f64) -> Color

Arguments:

r: Red value [0..1]

g: Green value [0..1]

b: Blue value [0..1]

a: Alpha value [0..1]

pub fn from_rgb_u8(r: u8, g: u8, b: u8) -> Color

Arguments:

r: Red value [0..255]

g: Green value [0..255]

b: Blue value [0..255]

pub fn from_rgba_u8(r: u8, g: u8, b: u8, a: u8) -> Color

Arguments:

r: Red value [0..255]

g: Green value [0..255]

b: Blue value [0..255]

a: Alpha value [0..255]

pub fn from_linear_rgb(r: f64, g: f64, b: f64) -> Color

Arguments:

r: Red value [0..1]

g: Green value [0..1]

b: Blue value [0..1]

pub fn from_lrgb(r: f64, g: f64, b: f64) -> Color

ðŸ‘Ž Deprecated

Arguments:

r: Red value [0..1]

g: Green value [0..1]

b: Blue value [0..1]

pub fn from_linear_rgba(r: f64, g: f64, b: f64, a: f64) -> Color

Arguments:

r: Red value [0..1]

g: Green value [0..1]

b: Blue value [0..1]

a: Alpha value [0..1]

pub fn from_lrgba(r: f64, g: f64, b: f64, a: f64) -> Color

ðŸ‘Ž Deprecated

Arguments:

r: Red value [0..1]

g: Green value [0..1]

b: Blue value [0..1]

a: Alpha value [0..1]

pub fn from_linear_rgb_u8(r: u8, g: u8, b: u8) -> Color

Arguments:

r: Red value [0..255]

g: Green value [0..255]

b: Blue value [0..255]

pub fn from_lrgb_u8(r: u8, g: u8, b: u8) -> Color

ðŸ‘Ž Deprecated

Arguments:

r: Red value [0..255]

g: Green value [0..255]

b: Blue value [0..255]

pub fn from_linear_rgba_u8(r: u8, g: u8, b: u8, a: u8) -> Color

Arguments:

r: Red value [0..255]

g: Green value [0..255]

b: Blue value [0..255]

a: Alpha value [0..255]

pub fn from_lrgba_u8(r: u8, g: u8, b: u8, a: u8) -> Color

ðŸ‘Ž Deprecated

Arguments:

r: Red value [0..255]

g: Green value [0..255]

b: Blue value [0..255]

a: Alpha value [0..255]

pub fn from_hsv(h: f64, s: f64, v: f64) -> Color

Arguments:

h: Hue angle [0..360]

s: Saturation [0..1]

v: Value [0..1]

pub fn from_hsva(h: f64, s: f64, v: f64, a: f64) -> Color

Arguments:

h: Hue angle [0..360]

s: Saturation [0..1]

v: Value [0..1]

a: Alpha [0..1]

pub fn from_hsl(h: f64, s: f64, l: f64) -> Color

Arguments:

h: Hue angle [0..360]

s: Saturation [0..1]

l: Lightness [0..1]

pub fn from_hsla(h: f64, s: f64, l: f64, a: f64) -> Color

Arguments:

h: Hue angle [0..360]

s: Saturation [0..1]

l: Lightness [0..1]

a: Alpha [0..1]

pub fn from_hwb(h: f64, w: f64, b: f64) -> Color

Arguments:

h: Hue angle [0..360]

w: Whiteness [0..1]

b: Blackness [0..1]

pub fn from_hwba(h: f64, w: f64, b: f64, a: f64) -> Color

Arguments:

h: Hue angle [0..360]

w: Whiteness [0..1]

b: Blackness [0..1]

a: Alpha [0..1]

pub fn from_oklab(l: f64, a: f64, b: f64) -> Color

Arguments:

l: Perceived lightness

a: How green/red the color is

b: How blue/yellow the color is

pub fn from_oklaba(l: f64, a: f64, b: f64, alpha: f64) -> Color

Arguments:

l: Perceived lightness

a: How green/red the color is

b: How blue/yellow the color is

alpha: Alpha [0..1]

pub fn from_html<S: AsRef<str>>(s: S) -> Result<Color, ParseColorError>

Create color from CSS color string.

Examples

use csscolorparser::Color;

let c = Color::from_html("rgb(255,0,0)")?;

assert_eq!(c.rgba(), (1.0, 0.0, 0.0, 1.0));
assert_eq!(c.rgba_u8(), (255, 0, 0, 255));
assert_eq!(c.to_hex_string(), "#ff0000");
assert_eq!(c.to_rgb_string(), "rgb(255,0,0)");

pub fn rgba(&self) -> (f64, f64, f64, f64)

Returns: (r, g, b, a)

Red, green, blue and alpha in the range [0..1]

pub fn rgba_u8(&self) -> (u8, u8, u8, u8)

Returns: (r, g, b, a)

Red, green, blue and alpha in the range [0..255]

pub fn red(&self) -> f64

ðŸ‘Ž Deprecated

Get the red value [0..1].

pub fn green(&self) -> f64

ðŸ‘Ž Deprecated

Get the green value [0..1].

pub fn blue(&self) -> f64

ðŸ‘Ž Deprecated

Get the blue value [0..1].

pub fn alpha(&self) -> f64

ðŸ‘Ž Deprecated

Get the alpha value [0..1].

pub fn to_hsva(&self) -> (f64, f64, f64, f64)

Returns: (h, s, v, a)

h: Hue angle [0..360]

s: Saturation [0..1]

v: Value [0..1]

a: Alpha [0..1]

pub fn to_hsla(&self) -> (f64, f64, f64, f64)

Returns: (h, s, l, a)

h: Hue angle [0..360]

s: Saturation [0..1]

l: Lightness [0..1]

a: Alpha [0..1]

pub fn to_hwba(&self) -> (f64, f64, f64, f64)

Returns: (h, w, b, a)

h: Hue angle [0..360]

w: Whiteness [0..1]

b: Blackness [0..1]

a: Alpha [0..1]

pub fn to_linear_rgba(&self) -> (f64, f64, f64, f64)

Returns: (r, g, b, a)

Red, green, blue and alpha in the range [0..1]

pub fn to_lrgba(&self) -> (f64, f64, f64, f64)

ðŸ‘Ž Deprecated

Returns: (r, g, b, a)

Red, green, blue and alpha in the range [0..1]

pub fn to_linear_rgba_u8(&self) -> (u8, u8, u8, u8)

Returns: (r, g, b, a)

Red, green, blue and alpha in the range [0..255]

pub fn to_lrgba_u8(&self) -> (u8, u8, u8, u8)

ðŸ‘Ž Deprecated

Returns: (r, g, b, a)

Red, green, blue and alpha in the range [0..255]

pub fn to_oklaba(&self) -> (f64, f64, f64, f64)

Returns: (l, a, b, alpha)

pub fn to_hex_string(&self) -> String

Get the RGB hexadecimal color string.

pub fn to_rgb_string(&self) -> String

Get the CSS rgb() format string.

pub fn interpolate_rgb(&self, other: &Color, t: f64) -> Color

Blend this color with the other one, in the RGB color-space. t in the range [0..1].

pub fn interpolate_linear_rgb(&self, other: &Color, t: f64) -> Color

Blend this color with the other one, in the linear RGB color-space. t in the range [0..1].

pub fn interpolate_lrgb(&self, other: &Color, t: f64) -> Color

ðŸ‘Ž Deprecated

Blend this color with the other one, in the linear RGB color-space. t in the range [0..1].

pub fn interpolate_hsv(&self, other: &Color, t: f64) -> Color

Blend this color with the other one, in the HSV color-space. t in the range [0..1].

pub fn interpolate_oklab(&self, other: &Color, t: f64) -> Color

Blend this color with the other one, in the Oklab color-space. t in the range [0..1].