Struct csscolorparser::Color

pub struct Color {
    pub r: f64,
    pub g: f64,
    pub b: f64,
    pub a: f64,
}

The color

Fields
r: f64

Red

g: f64

Green

b: f64

Blue

a: f64

Alpha

Implementations

impl Color

pub const fn new(r: f64, g: f64, b: f64, a: f64) -> Self

Arguments:

r: Red value [0..1]

g: Green value [0..1]

b: Blue value [0..1]

a: Alpha value [0..1]

pub fn to_array(&self) -> [f64; 4]

pub fn to_rgba8(&self) -> [u8; 4]

pub fn to_rgba16(&self) -> [u16; 4]

pub fn clamp(&self) -> Self

pub fn from_rgb(r: f64, g: f64, b: f64) -> Self

ğŸ‘Deprecated: Use new instead.

Arguments:

r: Red value [0..1]

g: Green value [0..1]

b: Blue value [0..1]

pub fn from_rgba(r: f64, g: f64, b: f64, a: f64) -> Self

ğŸ‘Deprecated: Use new instead.

Arguments:

r: Red value [0..1]

g: Green value [0..1]

b: Blue value [0..1]

a: Alpha value [0..1]

pub fn from_rgb_u8(r: u8, g: u8, b: u8) -> Self

ğŸ‘Deprecated: Use from_rgba8 instead.

Arguments:

r: Red value [0..255]

g: Green value [0..255]

b: Blue value [0..255]

pub fn from_rgba_u8(r: u8, g: u8, b: u8, a: u8) -> Self

ğŸ‘Deprecated: Use from_rgba8 instead.

Arguments:

r: Red value [0..255]

g: Green value [0..255]

b: Blue value [0..255]

a: Alpha value [0..255]

pub fn from_rgba8(r: u8, g: u8, b: u8, a: u8) -> Self

Arguments:

r: Red value [0..255]

g: Green value [0..255]

b: Blue value [0..255]

a: Alpha value [0..255]

pub fn from_linear_rgb(r: f64, g: f64, b: f64) -> Self

ğŸ‘Deprecated: Use from_linear_rgba instead.

Arguments:

r: Red value [0..1]

g: Green value [0..1]

b: Blue value [0..1]

pub fn from_linear_rgba(r: f64, g: f64, b: f64, a: f64) -> Self

Arguments:

r: Red value [0..1]

g: Green value [0..1]

b: Blue value [0..1]

a: Alpha value [0..1]

pub fn from_linear_rgb_u8(r: u8, g: u8, b: u8) -> Self

ğŸ‘Deprecated: Use from_linear_rgba8 instead.

Arguments:

r: Red value [0..255]

g: Green value [0..255]

b: Blue value [0..255]

pub fn from_linear_rgba_u8(r: u8, g: u8, b: u8, a: u8) -> Self

ğŸ‘Deprecated: Use from_linear_rgba8 instead.

Arguments:

r: Red value [0..255]

g: Green value [0..255]

b: Blue value [0..255]

a: Alpha value [0..255]

pub fn from_linear_rgba8(r: u8, g: u8, b: u8, a: u8) -> Self

Arguments:

r: Red value [0..255]

g: Green value [0..255]

b: Blue value [0..255]

a: Alpha value [0..255]

pub fn from_hsv(h: f64, s: f64, v: f64) -> Self

ğŸ‘Deprecated: Use from_hsva instead.

Arguments:

h: Hue angle [0..360]

s: Saturation [0..1]

v: Value [0..1]

pub fn from_hsva(h: f64, s: f64, v: f64, a: f64) -> Self

Arguments:

h: Hue angle [0..360]

s: Saturation [0..1]

v: Value [0..1]

a: Alpha [0..1]

pub fn from_hsl(h: f64, s: f64, l: f64) -> Self

ğŸ‘Deprecated: Use from_hsla instead.

Arguments:

h: Hue angle [0..360]

s: Saturation [0..1]

l: Lightness [0..1]

pub fn from_hsla(h: f64, s: f64, l: f64, a: f64) -> Self

Arguments:

h: Hue angle [0..360]

s: Saturation [0..1]

l: Lightness [0..1]

a: Alpha [0..1]

pub fn from_hwb(h: f64, w: f64, b: f64) -> Self

ğŸ‘Deprecated: Use from_hwba instead.

Arguments:

h: Hue angle [0..360]

w: Whiteness [0..1]

b: Blackness [0..1]

pub fn from_hwba(h: f64, w: f64, b: f64, a: f64) -> Self

Arguments:

h: Hue angle [0..360]

w: Whiteness [0..1]

b: Blackness [0..1]

a: Alpha [0..1]

pub fn from_oklab(l: f64, a: f64, b: f64) -> Self

ğŸ‘Deprecated: Use from_oklaba instead.

Arguments:

l: Perceived lightness

a: How green/red the color is

b: How blue/yellow the color is

pub fn from_oklaba(l: f64, a: f64, b: f64, alpha: f64) -> Self

Arguments:

l: Perceived lightness

a: How green/red the color is

b: How blue/yellow the color is

alpha: Alpha [0..1]

pub fn from_html<S: AsRef<str>>(s: S) -> Result<Self, ParseColorError>

Create color from CSS color string.

Examples

use csscolorparser::Color;

let c = Color::from_html("rgb(255,0,0)")?;

assert_eq!(c.to_array(), [1.0, 0.0, 0.0, 1.0]);
assert_eq!(c.to_rgba8(), [255, 0, 0, 255]);
assert_eq!(c.to_hex_string(), "#ff0000");
assert_eq!(c.to_rgb_string(), "rgb(255,0,0)");

pub fn name(&self) -> Option<&'static str>

pub fn rgba(&self) -> (f64, f64, f64, f64)

ğŸ‘Deprecated

Returns: (r, g, b, a)

Red, green, blue and alpha in the range [0..1]

pub fn rgba_u8(&self) -> (u8, u8, u8, u8)

ğŸ‘Deprecated: Use to_rgba8 instead.

Returns: (r, g, b, a)

Red, green, blue and alpha in the range [0..255]

pub fn to_hsva(&self) -> (f64, f64, f64, f64)

Returns: (h, s, v, a)

h: Hue angle [0..360]

s: Saturation [0..1]

v: Value [0..1]

a: Alpha [0..1]

pub fn to_hsla(&self) -> (f64, f64, f64, f64)

Returns: (h, s, l, a)

h: Hue angle [0..360]

s: Saturation [0..1]

l: Lightness [0..1]

a: Alpha [0..1]

pub fn to_hwba(&self) -> (f64, f64, f64, f64)

Returns: (h, w, b, a)

h: Hue angle [0..360]

w: Whiteness [0..1]

b: Blackness [0..1]

a: Alpha [0..1]

pub fn to_linear_rgba(&self) -> (f64, f64, f64, f64)

Returns: (r, g, b, a)

Red, green, blue and alpha in the range [0..1]

pub fn to_linear_rgba_u8(&self) -> (u8, u8, u8, u8)

Returns: (r, g, b, a)

Red, green, blue and alpha in the range [0..255]

pub fn to_oklaba(&self) -> (f64, f64, f64, f64)

Returns: (l, a, b, alpha)

pub fn to_hex_string(&self) -> String

Get the RGB hexadecimal color string.

pub fn to_rgb_string(&self) -> String

Get the CSS rgb() format string.

pub fn interpolate_rgb(&self, other: &Color, t: f64) -> Self

Blend this color with the other one, in the RGB color-space. t in the range [0..1].

pub fn interpolate_linear_rgb(&self, other: &Color, t: f64) -> Self

Blend this color with the other one, in the linear RGB color-space. t in the range [0..1].

pub fn interpolate_hsv(&self, other: &Color, t: f64) -> Self

Blend this color with the other one, in the HSV color-space. t in the range [0..1].

pub fn interpolate_oklab(&self, other: &Color, t: f64) -> Self

Blend this color with the other one, in the Oklab color-space. t in the range [0..1].